# perf: Drop use of Into::into for '?'