--- a/src/builder/value_parser.rs
+++ b/src/builder/value_parser.rs
@@ -569,7 +569,7 @@ where
         arg: Option<&crate::Arg>,
         value: &std::ffi::OsStr,
     ) -> Result<AnyValue, crate::Error> {
-        let value = TypedValueParser::parse_ref(self, cmd, arg, value)?;
+        let value = ok!(TypedValueParser::parse_ref(self, cmd, arg, value));
         Ok(AnyValue::new(value))
     }
 
@@ -579,7 +579,7 @@ where
         arg: Option<&crate::Arg>,
         value: std::ffi::OsString,
     ) -> Result<AnyValue, crate::Error> {
-        let value = TypedValueParser::parse(self, cmd, arg, value)?;
+        let value = ok!(TypedValueParser::parse(self, cmd, arg, value));
         Ok(AnyValue::new(value))
     }
 
@@ -694,18 +694,18 @@ where
         arg: Option<&crate::Arg>,
         value: &std::ffi::OsStr,
     ) -> Result<Self::Value, crate::Error> {
-        let value = value.to_str().ok_or_else(|| {
+        let value = ok!(value.to_str().ok_or_else(|| {
             crate::Error::invalid_utf8(
                 cmd,
                 crate::output::Usage::new(cmd).create_usage_with_title(&[]),
             )
-        })?;
-        let value = (self)(value).map_err(|e| {
+        }));
+        let value = ok!((self)(value).map_err(|e| {
             let arg = arg
                 .map(|a| a.to_string())
                 .unwrap_or_else(|| "...".to_owned());
             crate::Error::value_validation(arg, value.to_owned(), e.into()).with_cmd(cmd)
-        })?;
+        }));
         Ok(value)
     }
 }
@@ -742,12 +742,12 @@ impl TypedValueParser for StringValueParser {
         _arg: Option<&crate::Arg>,
         value: std::ffi::OsString,
     ) -> Result<Self::Value, crate::Error> {
-        let value = value.into_string().map_err(|_| {
+        let value = ok!(value.into_string().map_err(|_| {
             crate::Error::invalid_utf8(
                 cmd,
                 crate::output::Usage::new(cmd).create_usage_with_title(&[]),
             )
-        })?;
+        }));
         Ok(value)
     }
 }
@@ -939,7 +939,7 @@ impl<E: crate::ValueEnum + Clone + Send + Sync + 'static> TypedValueParser for E
                 .collect::<Vec<_>>()
         };
 
-        let value = value.to_str().ok_or_else(|| {
+        let value = ok!(value.to_str().ok_or_else(|| {
             crate::Error::invalid_value(
                 cmd,
                 value.to_string_lossy().into_owned(),
@@ -947,8 +947,8 @@ impl<E: crate::ValueEnum + Clone + Send + Sync + 'static> TypedValueParser for E
                 arg.map(ToString::to_string)
                     .unwrap_or_else(|| "...".to_owned()),
             )
-        })?;
-        let value = E::value_variants()
+        }));
+        let value = ok!(E::value_variants()
             .iter()
             .find(|v| {
                 v.to_possible_value()
@@ -963,7 +963,7 @@ impl<E: crate::ValueEnum + Clone + Send + Sync + 'static> TypedValueParser for E
                 arg.map(ToString::to_string)
                     .unwrap_or_else(|| "...".to_owned()),
             )
-            })?
+            }))
             .clone();
         Ok(value)
     }
@@ -1049,12 +1049,12 @@ impl TypedValueParser for PossibleValuesParser {
         arg: Option<&crate::Arg>,
         value: std::ffi::OsString,
     ) -> Result<String, crate::Error> {
-        let value = value.into_string().map_err(|_| {
+        let value = ok!(value.into_string().map_err(|_| {
             crate::Error::invalid_utf8(
                 cmd,
                 crate::output::Usage::new(cmd).create_usage_with_title(&[]),
             )
-        })?;
+        }));
 
         let ignore_case = arg.map(|a| a.is_ignore_case_set()).unwrap_or(false);
         if self.0.iter().any(|v| v.matches(&value, ignore_case)) {
@@ -1233,13 +1233,13 @@ where
         arg: Option<&crate::Arg>,
         raw_value: &std::ffi::OsStr,
     ) -> Result<Self::Value, crate::Error> {
-        let value = raw_value.to_str().ok_or_else(|| {
+        let value = ok!(raw_value.to_str().ok_or_else(|| {
             crate::Error::invalid_utf8(
                 cmd,
                 crate::output::Usage::new(cmd).create_usage_with_title(&[]),
             )
-        })?;
-        let value = value.parse::<i64>().map_err(|err| {
+        }));
+        let value = ok!(value.parse::<i64>().map_err(|err| {
             let arg = arg
                 .map(|a| a.to_string())
                 .unwrap_or_else(|| "...".to_owned());
@@ -1249,7 +1249,7 @@ where
                 err.into(),
             )
             .with_cmd(cmd)
-        })?;
+        }));
         if !self.bounds.contains(&value) {
             let arg = arg
                 .map(|a| a.to_string())
@@ -1263,7 +1263,7 @@ where
         }
 
         let value: Result<Self::Value, _> = value.try_into();
-        let value = value.map_err(|err| {
+        let value = ok!(value.map_err(|err| {
             let arg = arg
                 .map(|a| a.to_string())
                 .unwrap_or_else(|| "...".to_owned());
@@ -1273,7 +1273,7 @@ where
                 err.into(),
             )
             .with_cmd(cmd)
-        })?;
+        }));
 
         Ok(value)
     }
@@ -1431,13 +1431,13 @@ where
         arg: Option<&crate::Arg>,
         raw_value: &std::ffi::OsStr,
     ) -> Result<Self::Value, crate::Error> {
-        let value = raw_value.to_str().ok_or_else(|| {
+        let value = ok!(raw_value.to_str().ok_or_else(|| {
             crate::Error::invalid_utf8(
                 cmd,
                 crate::output::Usage::new(cmd).create_usage_with_title(&[]),
             )
-        })?;
-        let value = value.parse::<u64>().map_err(|err| {
+        }));
+        let value = ok!(value.parse::<u64>().map_err(|err| {
             let arg = arg
                 .map(|a| a.to_string())
                 .unwrap_or_else(|| "...".to_owned());
@@ -1447,7 +1447,7 @@ where
                 err.into(),
             )
             .with_cmd(cmd)
-        })?;
+        }));
         if !self.bounds.contains(&value) {
             let arg = arg
                 .map(|a| a.to_string())
@@ -1461,7 +1461,7 @@ where
         }
 
         let value: Result<Self::Value, _> = value.try_into();
-        let value = value.map_err(|err| {
+        let value = ok!(value.map_err(|err| {
             let arg = arg
                 .map(|a| a.to_string())
                 .unwrap_or_else(|| "...".to_owned());
@@ -1471,7 +1471,7 @@ where
                 err.into(),
             )
             .with_cmd(cmd)
-        })?;
+        }));
 
         Ok(value)
     }
@@ -1622,12 +1622,12 @@ impl TypedValueParser for FalseyValueParser {
         _arg: Option<&crate::Arg>,
         value: &std::ffi::OsStr,
     ) -> Result<Self::Value, crate::Error> {
-        let value = value.to_str().ok_or_else(|| {
+        let value = ok!(value.to_str().ok_or_else(|| {
             crate::Error::invalid_utf8(
                 cmd,
                 crate::output::Usage::new(cmd).create_usage_with_title(&[]),
             )
-        })?;
+        }));
         let value = if value.is_empty() {
             false
         } else {
@@ -1719,19 +1719,19 @@ impl TypedValueParser for BoolishValueParser {
         arg: Option<&crate::Arg>,
         value: &std::ffi::OsStr,
     ) -> Result<Self::Value, crate::Error> {
-        let value = value.to_str().ok_or_else(|| {
+        let value = ok!(value.to_str().ok_or_else(|| {
             crate::Error::invalid_utf8(
                 cmd,
                 crate::output::Usage::new(cmd).create_usage_with_title(&[]),
             )
-        })?;
-        let value = crate::util::str_to_bool(value).ok_or_else(|| {
+        }));
+        let value = ok!(crate::util::str_to_bool(value).ok_or_else(|| {
             let arg = arg
                 .map(|a| a.to_string())
                 .unwrap_or_else(|| "...".to_owned());
             crate::Error::value_validation(arg, value.to_owned(), "value was not a boolean".into())
                 .with_cmd(cmd)
-        })?;
+        }));
         Ok(value)
     }
 
@@ -1808,12 +1808,12 @@ impl TypedValueParser for NonEmptyStringValueParser {
                     .unwrap_or_else(|| "...".to_owned()),
             ));
         }
-        let value = value.to_str().ok_or_else(|| {
+        let value = ok!(value.to_str().ok_or_else(|| {
             crate::Error::invalid_utf8(
                 cmd,
                 crate::output::Usage::new(cmd).create_usage_with_title(&[]),
             )
-        })?;
+        }));
         Ok(value.to_owned())
     }
 }
@@ -1860,7 +1860,7 @@ where
         arg: Option<&crate::Arg>,
         value: &std::ffi::OsStr,
     ) -> Result<Self::Value, crate::Error> {
-        let value = self.parser.parse_ref(cmd, arg, value)?;
+        let value = ok!(self.parser.parse_ref(cmd, arg, value));
         let value = (self.func)(value);
         Ok(value)
     }
@@ -1871,7 +1871,7 @@ where
         arg: Option<&crate::Arg>,
         value: std::ffi::OsString,
     ) -> Result<Self::Value, crate::Error> {
-        let value = self.parser.parse(cmd, arg, value)?;
+        let value = ok!(self.parser.parse(cmd, arg, value));
         let value = (self.func)(value);
         Ok(value)
     }