--- a/src/derive.rs
+++ b/src/derive.rs
@@ -94,7 +94,7 @@ pub trait Parser: FromArgMatches + CommandFactory + Sized {
 
     /// Parse from `std::env::args_os()`, return Err on error.
     fn try_parse() -> Result<Self, Error> {
-        let mut matches = <Self as CommandFactory>::command().try_get_matches()?;
+        let mut matches = ok!(<Self as CommandFactory>::command().try_get_matches());
         <Self as FromArgMatches>::from_arg_matches_mut(&mut matches).map_err(format_error::<Self>)
     }
 
@@ -123,7 +123,7 @@ pub trait Parser: FromArgMatches + CommandFactory + Sized {
         I: IntoIterator<Item = T>,
         T: Into<OsString> + Clone,
     {
-        let mut matches = <Self as CommandFactory>::command().try_get_matches_from(itr)?;
+        let mut matches = ok!(<Self as CommandFactory>::command().try_get_matches_from(itr));
         <Self as FromArgMatches>::from_arg_matches_mut(&mut matches).map_err(format_error::<Self>)
     }
 
@@ -150,7 +150,7 @@ pub trait Parser: FromArgMatches + CommandFactory + Sized {
         T: Into<OsString> + Clone,
     {
         let mut matches =
-            <Self as CommandFactory>::command_for_update().try_get_matches_from(itr)?;
+            ok!(<Self as CommandFactory>::command_for_update().try_get_matches_from(itr));
         <Self as FromArgMatches>::update_from_arg_matches_mut(self, &mut matches)
             .map_err(format_error::<Self>)
     }