--- a/src/macros.rs
+++ b/src/macros.rs
@@ -664,12 +664,22 @@ macro_rules! debug {
 
 macro_rules! ok {
     ($expr:expr) => {
-        $expr?
+        match $expr {
+            Ok(val) => val,
+            Err(err) => {
+                return Err(err);
+            }
+        }
     };
 }
 
 macro_rules! some {
     ($expr:expr) => {
-        $expr?
+        match $expr {
+            Some(val) => val,
+            None => {
+                return None;
+            }
+        }
     };
 }