--- a/src/util/flat_map.rs
+++ b/src/util/flat_map.rs
@@ -65,11 +65,11 @@ impl<K: PartialEq + Eq, V> FlatMap<K, V> {
         K: Borrow<Q>,
         Q: std::hash::Hash + Eq,
     {
-        let index = self
+        let index = some!(self
             .keys
             .iter()
             .enumerate()
-            .find_map(|(i, k)| (k.borrow() == key).then(|| i))?;
+            .find_map(|(i, k)| (k.borrow() == key).then(|| i)));
         let key = self.keys.remove(index);
         let value = self.values.remove(index);
         Some((key, value))